# perf: replace cidranger with bart for significant performance improvements