Replace `cidranger` with `bart` for IP range checking in the `RemoteAddrChecker`, delivering significant performance improvements with zero heap allocations.

## Key Changes
- Replace `cidranger.Ranger` with `bart.Lite` in `RemoteAddrChecker` struct
- Update to modern `netip` package instead of legacy `net` package for IP handling
- Remove cidranger dependency from go.mod

## Performance Results

 < /dev/null |  Operation | cidranger | bart | Improvement |
|-----------|-----------|------|-------------|
| IPv4 lookups | 63.25ns, 20B, 2 allocs | 15.58ns, 0B, 0 allocs | **4x faster, 0 allocs** |
| IPv6 lookups | 76.96ns, 32B, 2 allocs | 26.51ns, 0B, 0 allocs | **3x faster, 0 allocs** |
| Insertions | 19,191ns, 14,080B, 644 allocs | 976ns, 784B, 16 allocs | **20x faster** |
| Large table lookups | 74.85ns, 4B, 1 alloc | 5.196ns, 0B, 0 allocs | **14x faster, 0 allocs** |

All existing functionality preserved with improved type safety using modern `netip` types.

Signed-off-by: Jason Cameron <git@jasoncameron.dev>

Checklist:

- [x] Added a description of the changes to the `[Unreleased]` section of docs/docs/CHANGELOG.md
- [x] Added test cases to [the relevant parts of the codebase](https://anubis.techaro.lol/docs/developer/code-quality)
- [x] Ran integration tests `npm run test:integration` (unsupported on Windows, please use WSL)