@thess 2025-06-17
Nice addition -- When you merge this I will update #593 as well as using FastHash.